Merged in feature/12-deduplication (pull request #14)
Correct query in finding duplicate records

The query in sync.StudyMethods.query_record() was incorrect. It did not limit the match within a single provenance item, but included matches from all items in the provenance list. Further, it included the direct attribute which is not needed when looking for matches.

The query is corrected. It uses $elemMatch to match the filter against a single provenance item, and loops thought every provenance item in new record when querying for matches.

Tests changed accordingly.

Add fixed item to 0.2.0 changelog entry.

Fixes #12 at Bitbucket.